requirements.txt
pytesseract==0.3.10
python-docx==0.8.11
PyPDF2==3.0.1
Pillow==10.0.0